explicit ``decimal.Decimal``
By explicitly serializing ``decimal.Decimal`` as type ``Decimal``, we
avoid rounding errors on deserialization as ``float``s. The alternative
would be to call `simplejson.loads` with ``parse_float=Decimal``, but
than all JSON decimal numbers are represented as Decimals -- and we lose
the distinction between (pre-encoding) ``float`` and ``Decimal``.

If the later is the wanted behaviour, call ``dumps`` with
``use_decimal=True`` (and ``loads`` with ``parse_float=decimal.Decimal``).