introduced ``exact`` and ``compat`` serialization
Should we aim for the *exact* reproduction of Python types,
or for maximum *compatibility* when (de-)serializing?

By default, we prefer the exactness of reproduction.
For example, `tuple`, `namedtuple`, `Decimal`, or `datetime` will all be
restored to the appropriate type (same as the starting type -- even the
custom class for the `namedtuple` is recreated).
When compatible coding if turned on, we shall fallback to standard JSON
types, and values from the example above will be serialized as
`list` (`Array`), `dict` (`Object`), `Number` and `ISO8601 timestamp string`,
respectively.

Please note:
  - `compat` mode is lossy -- `namedtuple` serialized as `dict`/`Object`
    can never be deserialized as `namedtuple`.
  - `exact` mode is verbose -- and if you have a standard JSON decoder  on
    the other end, all that additional type info is useless/discared.

To switch between representation styles, use  `jsonplus.prefer(coding)`,
where `coding` is `jsonplus.EXACT` or `jsonplus.COMPAT`. Another way,
maybe simpler, is to use `jsonplus.prefer_exact()` and
`jsonplus.prefer_compat()`.

The preference is stored thread-local.